requirements.txt
# local package
-e .

# external requirements
snakemake
click
Sphinx
coverage
#awscli
jupyterlab
black
isort
flake8
python-dotenv>=0.5.1
pandas
spacy
dask
jsonlines
spacyfishing
scrapy
scrapy-playwright

germansentiment